requirements.txt
aiogram
asyncpg
telethon
loguru
psycopg2-binary
tabulate
